Intial commit - skeleton and very basic functionality.
A MongoDB instance is required with a database 'agents'.
Two collections will be used for 'users' and 'posts', with simple pages to
display the contents of both.
A 'newpost' directive is available to also create new posts into the database.
There is currently no way to add users to the database from tha application.